+++ docs-src/angular-19/19.2.10/stackline-smoke.ts
@@ -1,0 +1,4 @@
+import * as StacklinePackage from '@stackline/ng-multiselect-dropdown';
+
+const packageExports = StacklinePackage;
+void packageExports;