+++ docs-src/angular-20/src/app/app.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
-import { NgMultiSelectDropDownModule } from '@revivejs/ng-multiselect-dropdown';
+import { NgMultiSelectDropDownModule } from '@stackline/ng-multiselect-dropdown';
 
 import { AppComponent } from './app.component';